requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0
fastapi>=0.110.0
uvicorn>=0.29.0
streamlit>=1.33.0
pydantic>=2.6.0
PyYAML>=6.0.0
joblib>=1.3.0
matplotlib>=3.8.0
plotly>=5.20.0
seaborn>=0.13.0
pyarrow>=15.0.0
pytest>=8.0.0
httpx>=0.27.0
